move getMovies() from baseViewModel to HomeViewModel
